split out context.test
